% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_qSIP_MAGs.R
\name{as.Num}
\alias{as.Num}
\title{The following function is adapted from HTSSIP R package
conversion to numeric}
\usage{
as.Num(x)
}
\arguments{
\item{x}{single value}
}
\value{
numeric
}
\description{
Conducts conversion: as.character --> as.numeric
}
