% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atomX-data.R
\docType{data}
\name{atomX}
\alias{atomX}
\title{Atom fraction excess table}
\format{
An object of class "list"
}
\usage{
data(atomX)
}
\description{
Data table generated from the "qSIP_atom_excess_MAGs" function
}
\keyword{datasets}
