% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_qSIP_MAGs.R
\name{calc_atom_excess_MAGs}
\alias{calc_atom_excess_MAGs}
\title{Calculate atom fraction excess}
\usage{
calc_atom_excess_MAGs(df_OTU_s, isotope = "13C")
}
\arguments{
\item{df_OTU_s}{OTU table with calculated AFE and delta BDs}

\item{isotope}{Isotopic substrate used in the experiment}
}
\value{
numeric value: atom fraction excess (A)
}
\description{
See Hungate et al., 2015 for more details
}
