% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_qSIP_MAGs.R
\name{calc_Mheavymax_MAGs}
\alias{calc_Mheavymax_MAGs}
\title{Calculate Mheavymax}
\usage{
calc_Mheavymax_MAGs(isotope = "13C", df_OTU_s1)
}
\arguments{
\item{isotope}{The isotope for which the DNA is labeled with ('13C' or '18O')}

\item{Mlight}{The molecular wight of unlabeled DNA}

\item{Gi}{The G+C content of unlabeled DNA}
}
\value{
numeric value: maximum molecular weight of fully-labeled DNA
}
\description{
This script was adapted from https://github.com/buckleylab/HTSSIP/blob/master/R/qSIP_atom_excess.R
for use with genome-centric metagenomics. See Hungate et al., 2015 for more details
}
