% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df-boot-atomX-data.R
\docType{data}
\name{df_atomX_boot}
\alias{df_atomX_boot}
\title{Bootstrapped atom fraction excess table}
\format{
An object of class "data.frame"
}
\usage{
data(df_atomX_boot)
}
\description{
Data table generated from bostrapping the AFE table using the "qSIP_bootstrap_fcr" function
}
\keyword{datasets}
