% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage-data.R
\docType{data}
\name{f_tibble}
\alias{f_tibble}
\title{Coverage table}
\format{
An object of class "data.frame"
}
\usage{
data(f_tibble)
}
\description{
Coverage data used for many functions in the package
}
\keyword{datasets}
