% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdr_adjusted_pvalues.R
\name{fdr_multiple_hypothesis_testing}
\alias{fdr_multiple_hypothesis_testing}
\title{False detection rate (FDR) adjusted p-values}
\usage{
fdr_multiple_hypothesis_testing(atomX, alpha)
}
\arguments{
\item{atomX}{A list object created by \code{qSIP_atom_excess_MAGs()}}

\item{alpha}{Numerical value - Significance level for hypothesis testing}
}
\value{
A data.frame of FDR adjusted p-values for each MAG and a BOOLEAN
of whether or not the MAG is an incorporator based on a set significance level
}
\description{
Welch's  two sample t-tests are used to calculate p-values and FDR correction
is done using the "fdrtool" package
}
