% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fractions-data.R
\docType{data}
\name{fractions}
\alias{fractions}
\title{Fractions table}
\format{
An object of class "data.frame"
}
\usage{
data(fractions)
}
\description{
Fractions data used for many functions in the package
}
\keyword{datasets}
