% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mag-table-data.R
\docType{data}
\name{mag.table}
\alias{mag.table}
\title{MAG abundance table in phyloseq format}
\format{
An object of class "phyloseq"
}
\usage{
data(mag.table)
}
\description{
MAG abundances in the format of phyloseq object to be used in the qSIP and (MW-)HR-SIP pipeline
}
\keyword{datasets}
