% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master-phyloseq-data.R
\docType{data}
\name{phylo.qSIP}
\alias{phylo.qSIP}
\title{Master phyloseq object}
\format{
An object of class "phyloseq"
}
\usage{
data(phylo.qSIP)
}
\description{
Master phyloseq object
}
\keyword{datasets}
