% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_qSIP_MAGs.R
\name{phyloseq2table}
\alias{phyloseq2table}
\title{The following function is adapted from HTSSIP R package
Phyloseq conversion to a ggplot-formatted table}
\usage{
phyloseq2table(
  physeq,
  include_sample_data = FALSE,
  sample_col_keep = NULL,
  include_tax_table = FALSE,
  tax_col_keep = NULL,
  control_expr = NULL
)
}
\arguments{
\item{physeq}{Phyloseq object}

\item{include_sample_data}{Include \code{sample_table} information?}

\item{sample_col_keep}{Which columns in the \code{sample_data} table to keep?
Use \code{NULL} to keep all columns.}

\item{include_tax_table}{Include \code{tax_table} information?}

\item{tax_col_keep}{A vector for column names to keep.
Use \code{NULL} to keep all columns.}

\item{control_expr}{An expression for identifying which samples are controls.
Control/non-control identification will be in the 'IS_CONTROL' column of the
returned data.frame object.}
}
\value{
data.frame
}
\description{
Convert the OTU table (+ metadata) to a format that can be
easily plotted with phyloseq
}
