% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_qSIP_MAGs.R
\name{qSIP_atom_excess_format_MAGs}
\alias{qSIP_atom_excess_format_MAGs}
\title{Reformat a phyloseq object of qSIP_atom_excess_MAGs analysis}
\usage{
qSIP_atom_excess_format_MAGs(physeq, control_expr, treatment_rep)
}
\arguments{
\item{physeq}{A phyloseq object}

\item{control_expr}{An expression for identifying unlabeled control
samples in the phyloseq object (eg., "Substrate=='12C-Con'")}

\item{treatment_rep}{Which column in the phyloseq sample data designates
replicate treatments}
}
\value{
numeric value: atom fraction excess (A)
}
\description{
Reformat a phyloseq object of qSIP_atom_excess_MAGs analysis
}
