% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples-object-data.R
\docType{data}
\name{samples.object}
\alias{samples.object}
\title{Fractions table in phyloseq format}
\format{
An object of class "phyloseq"
}
\usage{
data(samples.object)
}
\description{
Fractions metadata in the format of phyloseq object to be used in the qSIP and (MW-)HR-SIP pipeline
}
\keyword{datasets}
