% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequin-dilutions-data.R
\docType{data}
\name{seq_dil}
\alias{seq_dil}
\title{Sequins dilution table}
\format{
An object of class "data.frame"
}
\usage{
data(seq_dil)
}
\description{
Sequins dilution data
}
\keyword{datasets}
