% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BD_abundance_viz.R
\name{sip_vis}
\alias{sip_vis}
\title{Create a table for easy data visualization of plots with
buoyant densities on X-axis and absolute abundance on Y-axis}
\usage{
sip_vis(fractions, mag_tab, taxonomy, afe_tab)
}
\arguments{
\item{mag_tab}{mag_tab: a tibble with first column "Feature" that contains the bin ID (or contig IDs),
and the rest of the columns represent samples with features' scaled abundances (attamoles/uL)}

\item{fractions_df}{A fractions file with the following columns
\itemize{
\item Replicate: Depends on how many replicates the study has
\item Fractions: Typically in the range of 2-24
\item Buoyant_density: As calculated from the refractometer for each fraction and replicate
\item Isotope: "12C", "13C", "14N", "15N" etc.
\item DNA_concentration
\item Sample: In the format "'isotope'\emph{rep}#\emph{fraction}#".
For instance, "12C_rep_1_fraction_1"
}}
}
\value{
data  frame: formatted table for BD vs absolute abundance visualization
}
\description{
Create a table for easy data visualization of plots with
buoyant densities on X-axis and absolute abundance on Y-axis
}
