% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy-object.R
\docType{data}
\name{taxonomy.object}
\alias{taxonomy.object}
\title{Taxonomy table in phyloseq format}
\format{
An object of class "phyloseq"
}
\usage{
data(taxonomy.object)
}
\description{
Taxonomy table in the format of phyloseq object to be used in the qSIP and (MW-)HR-SIP pipeline
}
\keyword{datasets}
