% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy-table-data.R
\docType{data}
\name{taxonomy_tibble}
\alias{taxonomy_tibble}
\title{Taxonomy table}
\format{
An object of class "data.frame"
}
\usage{
data(taxonomy_tibble)
}
\description{
Taxonomy table from GTDB-Tk output - combining both bacterial and archaeal taxonomy
}
\keyword{datasets}
